requirements.txt
flask
flask-compress
gunicorn
python-frontmatter
markdown
requests
Pillow
python-dotenv
google-genai
